delete y create en show con error resuelto de appedn

+++ resources/views/comment/qjs.blade.php
@@ -7,7 +7,8 @@
 
 /* ======== agregar ======== */
 $(function() {
-  $('#sendComment').on('click', function(e) {
+  $(document).on('click','#sendComment', function(e) {
+    e.preventDefault();
     let text = document.getElementById('MakeCommentSection').value;
     createComment(text)
   });
@@ -41,10 +42,6 @@ function createComment(text) {
 }
 
 
-
-
-
-
 /* ======== listar ======== */
 function allComents(){
 
@@ -55,7 +52,7 @@ $.ajax({
     url:url,
         success:function(data){
         console.log(data);
-         addFieldswithOptionalException(data['comments']['data'], data['userComments'], -1 );
+         addFieldswithOptionalException(data['comments']['data'], data['userComments'],data['commentsDiffForHuman'] ,-1 );
         },
         error:function(){
             console.log(JSON.stringify(error));
@@ -64,44 +61,77 @@ $.ajax({
 
 }
 
-function addFieldswithOptionalException(comments, user, exception) {
+function addFieldswithOptionalException(comments, userComment, diffForHumans,  exception) {
     console.log(comments);
-    $( "div" ).remove( "#commentsSection" );
+    $('.commentsSection').html('');
+    let commentSection="";
 
-  /*   for (let i = 0; i < comments.length; i++) {
-        var CommentSectionDiv
-        console.log(comments[i]['text']);
+    for (let i = 0; i < comments.length; i++) {
+         commentSection += "<div class='card border-dark mb-3' style='' id='commentCard'>"
+         commentSection += "<div class='card-body'>"
+         commentSection += " <div class='row'>"
+         commentSection += "<h4 class='card-title col-10 mr-5' id=''>"+userComment[i]['name'] +"</h4>"
 
-    } */
-
-}
-
-
+        if (userComment[i]['id'] == user['id']) {
+            commentSection += "<button class='deleteIcon2 btn col-auto ml-4 ' onClick='deleteQualification(this.id)' id='"+comments[i]['id'] +"'><i class='ml-1 fa-solid fa-trash' ></i> </button>"
+            commentSection += "<button class='updateIcon btn col-auto' id='"+comments[i]['id'] +"' ><i class='ml-auto fa-solid fa-pen-to-square' ></i> </button>"
 
 
-/* eliminar comentario */
- function deleteQualification(id) {
-     console.log(" a eliminar "+ id['id']);
-
-     $.ajaxSetup({
-        headers: {
-            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
         }
-    });
-
-     $.ajax({
-         type:"DELETE",
-         url:"/qualifications/"+id,
-            success:function(data){
-             console.log(data);
-
-            },
-            error:function(){
-                console.log(JSON.stringify(error));
-            }
-		});
-
+         commentSection += "<h6 class='card-subtitle mb-2 text-muted'>"+diffForHumans[i]+"</h6>"
+         commentSection += "<p class='card-text'>"+ comments[i]['text'] +"</p>"
+         commentSection += "</div></div></div>"
+         console.log(commentSection);
+         $('.commentsSection').append(commentSection);
+        }
 }
+
+
+/* ======== eliminar ======== */
+
+ function deleteQualification(id) {
+    var  url =   '{{route("comments.destroy", ":id") }}'
+    url = url.replace(':id', id)
+    Swal.fire({
+        title: '¿Estas seguro?',
+        text: "No podrás revertir esto!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'si, eliminalo!'
+        }).then((result) => {
+        if (result.isConfirmed) {
+
+            $.ajaxSetup({
+                headers: {
+                    'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+                }
+            });
+            $.ajax({
+                type:"DELETE",
+                data:{'property_id': property['id']},
+                url:url,
+                    success:function(data){
+                        console.log(data);
+                        allComents();
+                        Swal.fire(
+                            'Eliminado!',
+                            'Elemento elimnado.',
+                            'success'
+                        )
+
+                    },
+                    error:function(){
+                        console.log(JSON.stringify(error));
+                    }
+                });
+            }
+
+
+        });
+
+ }
 
 
 /* editar comentarios */
